+++ src/components/common/index.tsx
@@ -1,0 +1,264 @@
+import { motion } from 'framer-motion'
+import Loader from 'react-loader-spinner'
+import styled from 'styled-components'
+import { device } from 'utils/style/responsive'
+
+export const DPDown = styled(motion.div)``
+
+export const StyledLoader = styled(Loader)`
+  display: flex;
+  justify-content: center;
+  margin: 50px;
+`
+
+export const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  flex: 1;
+  width: 100%;
+  height: 100vh;
+`
+
+export const Main = styled.main`
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  align-items: center;
+`
+
+export const Form = styled(motion.form)`
+  display: flex;
+  flex-direction: column;
+  flex: 1 1 auto;
+  width: 100%;
+  align-items: center;
+`
+
+export const FormContent = styled.div`
+  display: flex;
+  flex-direction: column;
+  flex: 1 1 auto;
+  width: 100%;
+  align-items: center;
+  overflow-y: auto;
+
+  &::-webkit-scrollbar {
+    width: 10px;
+  }
+
+  /* Track */
+  &::-webkit-scrollbar-track {
+    background: white;
+  }
+
+  /* Handle */
+  &::-webkit-scrollbar-thumb {
+    background: #ccc;
+  }
+
+  /* Handle on hover */
+  &::-webkit-scrollbar-thumb:hover {
+    background: #555;
+  }
+`
+
+export const DatePickerWrap = styled.div`
+  position: relative;
+  flex: 1;
+  width: 100%;
+`
+
+export const InnerForm = styled.div`
+  width: 100%;
+  flex: 1;
+  max-width: 500px;
+  padding: 5rem 0;
+
+  @media ${device.tablet} {
+    padding: 2rem;
+  }
+`
+
+export const FormGroup = styled.div`
+  background-color: var(--grey);
+  display: flex;
+  flex-direction: column;
+  padding: 2rem;
+  border-radius: 5px;
+  margin-bottom: 1.5rem;
+
+  strong {
+    font-weight: bold;
+  }
+`
+
+export const FormInnerGroup = styled(motion.div)`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  background: #f9f9fa;
+  border-radius: 10px;
+
+  margin-top: 2rem;
+  width: 100%;
+
+  &:first-child {
+    margin-top: 0;
+  }
+`
+
+export const Label = styled.label`
+  display: block;
+  font-size: 1.4rem;
+  margin-bottom: 2rem;
+  color: black;
+  display: block;
+  width: 100%;
+`
+
+export const LabelQuestion = styled(motion.label)`
+  display: block;
+  font-size: 1.4rem;
+  color: black;
+  display: block;
+  width: 100%;
+
+  strong {
+    font-weight: 600;
+  }
+
+  @media ${device.tablet} {
+    font-size: 1.2rem;
+  }
+`
+
+export const Input = styled.input`
+  padding: 1.3rem 1.6rem;
+  border-radius: 10px;
+  border: none;
+  font-size: 1.6rem;
+  border: 1px solid #f1f1f2;
+  outline: 0;
+  display: block;
+  width: 100%;
+  font-size: 1.4rem;
+  line-height: 2rem;
+
+  &::placeholder {
+    color: #d0d0ce;
+  }
+
+  &:focus {
+    border: 1px solid #ccc;
+  }
+
+  &.react-datepicker-ignore-onclickoutside {
+    border-radius: 10px 10px 0 0 !important;
+  }
+`
+
+export const RadioButtonGroup = styled(motion.div)`
+  display: flex;
+  margin-top: 2rem;
+`
+
+export const RadioButton = styled.label`
+  position: relative;
+  padding: 1rem 2rem;
+  border-radius: 10px;
+  background: #f1f1f2;
+  color: #76787b;
+  font-weight: 600;
+  margin-right: 1rem;
+  padding-left: 3.5rem;
+  font-size: 1.6rem;
+  line-height: 2rem;
+
+  > input {
+    opacity: 0;
+    visibility: hidden;
+    position: absolute;
+  }
+
+  > div {
+    content: '';
+    display: block;
+    width: 16px;
+    height: 16px;
+    border-radius: 16px;
+    border: 1px solid #d6d6d3;
+    background: white;
+    position: absolute;
+    left: 10px;
+    top: 50%;
+    margin-top: -8px;
+  }
+
+  > div:before {
+    content: '';
+    display: block;
+    width: 8px;
+    height: 8px;
+    border-radius: 10px;
+    background: black;
+    position: absolute;
+    left: 3px;
+    top: 3px;
+    transform: scale(0);
+    transition: transform 0.15s;
+  }
+
+  > input:checked ~ div:before {
+    transform: scale(1);
+  }
+`
+
+export const FormFooter = styled.div`
+  width: 100%;
+  border-top: 1px solid #f1f1f2;
+  padding: 2rem 2.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`
+
+export const Button = styled.button`
+  background: var(--accent);
+  font-size: 1.6rem;
+  padding: 1.3rem 2rem;
+  flex: 1;
+  max-width: 200px;
+  border-radius: 10px;
+  font-weight: 600;
+  text-align: left;
+  display: flex;
+  align-items: center;
+
+  > svg {
+    margin-left: auto;
+  }
+`
+
+export const FormButtonGroup = styled.div`
+  flex-direction: row;
+`
+
+export const AcceptDeleteButton = styled(motion.div)`
+  background: red;
+  font-size: 1.6rem;
+  padding: 1.3rem 2rem;
+  margin: 1.2rem;
+  max-width: 200px;
+  border-radius: 10px;
+  font-weight: 600;
+  text-align: left;
+  display: inline-block;
+  align-items: center;
+  cursor: pointer;
+  color: white;
+
+  > svg {
+    margin-left: auto;
+  }
+`
