+++ src/cypress/integration/pages/newTrip.spec.ts
@@ -1,4 +1,4 @@
-import { submitForm } from 'support/submitForm'
+import { submitForm } from '../../support/submitForm'
 
 describe('New Trip Page', () => {
   beforeEach(() => {
