+++ src/utils/style/customStyles.ts
@@ -1,0 +1,11 @@
+export const customStyles = {
+  content: {
+    top: '45%',
+    left: '50%',
+    right: 'auto',
+    bottom: 'auto',
+    marginRight: '-50%',
+    transform: 'translate(-50%, -50%)',
+    width: '500px',
+  },
+} as const
